Refactoring sur l'affichage de l'offre form + zone commentaire offre
