+++ saas_frontend/app/components/Form/Offre/OffreForm.tsx
@@ -1,16 +1,17 @@
 "use client";
-import React, { useState } from 'react';
-import { Accordion, Button, Alert, Timeline, FileInput, Label } from 'flowbite-react';
+import React, { useState, useEffect } from 'react';
+import { Accordion, Button, Alert, Timeline, FileInput, Label, Card } from 'flowbite-react';
 import ExtrasForm from '@/app/components/Form/Offre/ExtrasForm';
 import ConditionsFinancieresForm from '@/app/components/Form/Offre/ConditionsFinancieresForm';
 import BudgetEstimatifForm from '@/app/components/Form/Offre/BudgetEstimatifForm';
 import DetailOffreForm from '@/app/components/Form/Offre/DetailOffreForm';
-import DonneesSupplementairesForm from '@/app/components/Form/Offre/DonneesSupplementairesForm';
 import FicheTechniqueArtisteForm from '@/app/components/Form/Offre/FicheTechniqueArtiste';
 import InfoAdditionnelAlert from '@/app/components/Alerte/InfoAdditionnelAlerte';
 import { apiPost } from '@/app/services/internalApiClients';
 import { HiInformationCircle } from "react-icons/hi";
 import { FormData } from '@/app/types/FormDataType';
+import SelectCheckbox from '@/app/components/SelectCheckbox';
+import { apiGet } from '@/app/services/internalApiClients';
 
 const OffreForm: React.FC = () => {
     const dateParDefaut = new Date().toISOString().split('T')[0] as string;
@@ -26,15 +27,13 @@ const OffreForm: React.FC = () => {
             placesMin: null,
             placesMax: null,
             nbArtistesConcernes: null,
-            nbInvitesConcernes: null,
-            liensPromotionnels: []
+            nbInvitesConcernes: null
         },
         extras: {
             descrExtras: null,
             coutExtras: null,
             exclusivite: null,
             exception: null,
-            ordrePassage: null,
             clausesConfidentialites: null
         },
         etatOffre: {
@@ -59,15 +58,17 @@ const OffreForm: React.FC = () => {
             besoinEclairage: null,
             besoinEquipements: null,
             besoinScene: null,
-            besoinSonorisation: null
+            besoinSonorisation: null,
+            ordrePassage: null,
+            liensPromotionnels: [],
+            artiste: [],
+            nbArtistes: 0
         },
         donneesSupplementaires: {
             reseau: [],
             nbReseaux: 0,
             genreMusical: [],
             nbGenresMusicaux: 0,
-            artiste: [],
-            nbArtistes: 0
         },
         utilisateur: {
             username : typeof window !== 'undefined' ? sessionStorage.getItem('username') : ""
@@ -82,6 +83,10 @@ const OffreForm: React.FC = () => {
     const [typeMessage, setTypeMessage] = useState<"success" | "error">("success");
     const [offreId, setOffreId] = useState("");
     const [description, setDescription] = useState("");
+    const [genresMusicaux, setGenresMusicaux] = useState<Array<{ nomGenreMusical: string }>>([]);
+    const [selectedGenres, setSelectedGenres] = useState<string[]>(formData.donneesSupplementaires.genreMusical);
+    const [reseaux, setReseaux] = useState<Array<{ nomReseau: string }>>([]);
+    const [selectedReseaux, setSelectedReseaux] = useState<string[]>(formData.donneesSupplementaires.reseau);
 
     const valideFormulaire = async (e: React.FormEvent) => {
         e.preventDefault();
@@ -100,6 +105,33 @@ const OffreForm: React.FC = () => {
         }
     };
 
+    useEffect(() => {
+        const fetchGenresMusicaux = async () => {
+            try {
+                const genres = await apiGet('/genres-musicaux');
+                const genresList = JSON.parse(genres.genres_musicaux);
+                setGenresMusicaux(genresList);
+            } catch (error) {
+                console.error("Erreur lors du chargement des genres musicaux :", error);
+            }
+        };
+
+        const fetchReseauUtilisateur = async () => {
+            try {
+                const username = typeof window !== 'undefined' ? sessionStorage.getItem('username') : "";
+                const data = { username };
+                const datasUser = await apiPost('/utilisateur', JSON.parse(JSON.stringify(data)));
+                const reseauxListe: Array<{ nomReseau: string }> = JSON.parse(datasUser.utilisateur)[0].reseaux;
+                setReseaux(reseauxListe);
+            } catch (error) {
+                console.error("Erreur lors du chargement des données utilisateurs :", error);
+            }
+        };
+
+        fetchGenresMusicaux();
+        fetchReseauUtilisateur();
+    }, []);
+
     const checkInformationsDeBase = () => {
         const {
             titleOffre,
@@ -112,8 +144,7 @@ const OffreForm: React.FC = () => {
             placesMin,
             placesMax,
             nbArtistesConcernes,
-            nbInvitesConcernes,
-            liensPromotionnels
+            nbInvitesConcernes
         } = formData.detailOffre;
         return !!(titleOffre && 
             deadLine && 
@@ -125,8 +156,7 @@ const OffreForm: React.FC = () => {
             placesMin != null && placesMin > 0 && 
             placesMax != null && placesMax > 0 &&
             nbArtistesConcernes != null && nbArtistesConcernes > 0 &&
-            nbInvitesConcernes != null && nbInvitesConcernes > 0 &&
-            liensPromotionnels.length > 0);
+            nbInvitesConcernes != null && nbInvitesConcernes > 0);
     };
 
     const checkExtras = () => {
@@ -135,7 +165,6 @@ const OffreForm: React.FC = () => {
             coutExtras,
             exclusivite,
             exception,
-            ordrePassage,
             clausesConfidentialites 
         } = formData.extras;
         return !!(descrExtras && 
@@ -143,7 +172,6 @@ const OffreForm: React.FC = () => {
             coutExtras > 0 && 
             exclusivite && 
             exception && 
-            ordrePassage && 
             clausesConfidentialites);
     };
 
@@ -175,9 +203,23 @@ const OffreForm: React.FC = () => {
             besoinEclairage,
             besoinEquipements,
             besoinScene,
-            besoinSonorisation
+            besoinSonorisation,
+            ordrePassage,
+            liensPromotionnels,
+            artiste,
+            nbArtistes
         } = formData.ficheTechniqueArtiste;
-        return !!(besoinBackline && besoinEclairage && besoinEquipements && besoinScene && besoinSonorisation);
+        return !!(
+            besoinBackline && 
+            besoinEclairage && 
+            besoinEquipements && 
+            besoinScene && 
+            besoinSonorisation && 
+            ordrePassage && 
+            artiste.length > 0 && 
+            liensPromotionnels.length > 0 && 
+            nbArtistes != null && nbArtistes > 0
+        );
     };
 
     const checkDonneesSupplementaires = () => {
@@ -185,16 +227,22 @@ const OffreForm: React.FC = () => {
             reseau,
             nbReseaux,
             genreMusical,
-            nbGenresMusicaux,
-            artiste,
-            nbArtistes
+            nbGenresMusicaux
         } = formData.donneesSupplementaires;
         return reseau.length > 0 && 
             nbReseaux != null && nbReseaux > 0 && 
             genreMusical.length > 0 && 
-            nbGenresMusicaux != null && nbGenresMusicaux > 0 && 
-            artiste.length > 0 && 
-            nbArtistes != null && nbArtistes > 0;
+            nbGenresMusicaux != null && nbGenresMusicaux > 0;
+    };
+
+    const onDonneesSupplementairesChange = (name: string, value: any) => {
+        setFormData((prevData) => ({
+            ...prevData,
+            donneesSupplementaires: {
+                ...prevData.donneesSupplementaires,
+                [name]: value
+            }
+        }));
     };
 
     const updateField = (
@@ -355,31 +403,47 @@ const OffreForm: React.FC = () => {
                                 />
                             </Accordion.Content>
                         </Accordion.Panel>
-
                         <Accordion.Panel>
-                            <Accordion.Title>Données Supplémentaires</Accordion.Title>
+                            <Accordion.Title>Données de publications</Accordion.Title>
                             <Accordion.Content className='p-0'>
-                                <DonneesSupplementairesForm
-                                    donneesSupplementaires={formData.donneesSupplementaires}
-                                    onDonneesSupplementairesChange={(name, value) =>
-                                        setFormData((prevData) => ({
-                                            ...prevData,
-                                            donneesSupplementaires: {
-                                                ...prevData.donneesSupplementaires,
-                                                [name]: value
-                                            }
-                                        }))
-                                    }
-                                />
-                            </Accordion.Content>
+                            <Card className="w-full shadow-none border-none">
+                                <div className="grid grid-cols-2 gap-3">
+                                    <div>
+                                        <SelectCheckbox
+                                            domaineSelection="Réseaux sur lesquels poster votre évènement :"
+                                            options={reseaux.map((reseau) => ({ label: reseau.nomReseau, value: reseau.nomReseau }))}
+                                            selectedValues={selectedReseaux}
+                                            onSelectionChange={(updatedReseaux) => {
+                                                setSelectedReseaux(updatedReseaux);
+                                                onDonneesSupplementairesChange("reseau", updatedReseaux);
+                                                onDonneesSupplementairesChange("nbReseaux", updatedReseaux.length);
+                                            }}
+                                        />
+                                    </div>
+
+                                    <div>
+                                        <SelectCheckbox
+                                            domaineSelection="Genres musicaux en lien avec votre évènement :"
+                                            options={genresMusicaux.map((genreMusical) => ({ label: genreMusical.nomGenreMusical, value: genreMusical.nomGenreMusical }))}
+                                            selectedValues={selectedGenres}
+                                            onSelectionChange={(updatedGenres) => {
+                                                setSelectedGenres(updatedGenres);
+                                                onDonneesSupplementairesChange("genreMusical", updatedGenres);
+                                                onDonneesSupplementairesChange("nbGenresMusicaux", updatedGenres.length);
+                                            }}
+                                        />
+                                    </div>
+                                </div>
+                            </Card>
+                        </Accordion.Content>
                         </Accordion.Panel>
                     </Accordion>
                     <div>
                         <Label htmlFor="file-upload-helper-text" value="Ajouter une image de référence" />
                         <FileInput
                             id="file-upload-helper-text"
-                            helperText="JPEG or JPG"
-                            accept=".jpg,.jpeg"
+                            helperText="JPG"
+                            accept=".jpg"
                             onChange={(e) => {
                                 const file = e.target.files?.[0] || null;
                                 updateFieldWithBlob("image", "file", file);
