+++ saas_frontend/app/navigation/Router.tsx
@@ -10,7 +10,7 @@ const NavigationHandler: React.FC<NavigationHandlerProps> = ({ children }) => {
   const router = useRouter();
 
   const handleNavigation = async (path: string) => {
-    console.log('Navigating to' + path);
+    ('Navigating to' + path);
     // Précharge la page sans changer immédiatement
     // await router.prefetch(path); await inutil ???
